Added support for the Temperature and Humidity Sensor and the Temperature and Humidity Sensor Pro
Moved the existing implementation from class
TemperatureAndHumiditySensor to class TemperatureSensor. It contains the
support for the temperature sensors (without humidity) in versions 1.0,
1.1 and 1.2.

Adapted the class TemperatureAndHumiditySensor to support the two
sensors Temperature and Humidity Sensor (DHT11) and Temperature and
Humidity Sensor Pro (DHT22).

There are also other Temperature and Humidity Sensors like the
Temperature and Humidity Sensor (High-accuracy & mini) 1.0 which is
currently not supported as I don't have one for testing.